Advanced in implementing the deletion
Implemented a quarter of the part where the node to delete and the nodo to replace with are both black.
Specifically when the sibiling of the node to delete has a left sibiling black and has a red child.
not tested yet.
This is the symmetric case of the previous commit.