Test working for tree with dimension up to 11 nodes.
Fixed documentation.
Delete use getRoot for returning always the root of the three. I wrote that is discuraged to call delete on a subtree because the top parent three is returned.
Fixed priority of rebalance methods in fixdelete.
Fixed correct double linked pointer between parent and child in deletion.
Set the correct rotation in a case of fix deletion.