Removed bad designed RootDeletionException and created the code for avoid that case
Exceptions are for unexpected behaviour. If you are deleting the last element of the three you should expect that bahaviur.
Supporting empty tree would require code redisign, expecially around nullLeaf concept. In addition it doesn't have sense to me to have an empty data structure. Some case could require it but an encapsulation of it could be more appropriate.
Tested.